Move to test_project.py
