+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
+.DS_Store
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
